Replace ResourcePath with ResPath
